% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-cumulative-functions.R
\name{ckurtosis}
\alias{ckurtosis}
\title{Cumulative Kurtosis}
\usage{
ckurtosis(.x)
}
\arguments{
\item{.x}{A numeric vector}
}
\value{
A numeric vector
}
\description{
A function to return the cumulative kurtosis of a vector.
}
\details{
A function to return the cumulative kurtosis of a vector.
}
\examples{
x <- mtcars$mpg

ckurtosis(x)

}
\seealso{
Other Vector Function:
\code{\link[=cgmean]{cgmean()}},
\code{\link[=chmean]{chmean()}},
\code{\link[=cmean]{cmean()}},
\code{\link[=cmedian]{cmedian()}},
\code{\link[=crange]{crange()}},
\code{\link[=csd]{csd()}},
\code{\link[=cskewness]{cskewness()}},
\code{\link[=cvar]{cvar()}},
\code{\link[=euclidean_distance]{euclidean_distance()}},
\code{\link[=kurtosis_vec]{kurtosis_vec()}},
\code{\link[=rw_range]{rw_range()}},
\code{\link[=skewness_vec]{skewness_vec()}}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Vector Function}
