% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-cumulative-functions.R
\name{crange}
\alias{crange}
\title{Cumulative Range}
\usage{
crange(.x)
}
\arguments{
\item{.x}{A numeric vector}
}
\value{
A numeric vector
}
\description{
A function to return the cumulative range of a vector.
}
\details{
A function to return the cumulative range of a vector. It uses \code{max(.x[1:k]) - min(.x[1:k])} as
the basis of the function.
}
\examples{

x <- mtcars$mpg

crange(x)

}
\seealso{
Other Vector Function:
\code{\link[=cgmean]{cgmean()}},
\code{\link[=chmean]{chmean()}},
\code{\link[=ckurtosis]{ckurtosis()}},
\code{\link[=cmean]{cmean()}},
\code{\link[=cmedian]{cmedian()}},
\code{\link[=csd]{csd()}},
\code{\link[=cskewness]{cskewness()}},
\code{\link[=cvar]{cvar()}},
\code{\link[=euclidean_distance]{euclidean_distance()}},
\code{\link[=kurtosis_vec]{kurtosis_vec()}},
\code{\link[=rw_range]{rw_range()}},
\code{\link[=skewness_vec]{skewness_vec()}}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Vector Function}
