% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-cumulative-functions.R
\name{csd}
\alias{csd}
\title{Cumulative Standard Deviation}
\usage{
csd(.x)
}
\arguments{
\item{.x}{A numeric vector}
}
\value{
A numeric vector. Note: The first entry will always be
NaN.
}
\description{
A function to return the cumulative standard deviation of a vector.
}
\details{
A function to return the cumulative standard deviation of a vector.
}
\examples{
x <- mtcars$mpg

csd(x)

}
\seealso{
Other Vector Function:
\code{\link[=cgmean]{cgmean()}},
\code{\link[=chmean]{chmean()}},
\code{\link[=ckurtosis]{ckurtosis()}},
\code{\link[=cmean]{cmean()}},
\code{\link[=cmedian]{cmedian()}},
\code{\link[=crange]{crange()}},
\code{\link[=cskewness]{cskewness()}},
\code{\link[=cvar]{cvar()}},
\code{\link[=euclidean_distance]{euclidean_distance()}},
\code{\link[=kurtosis_vec]{kurtosis_vec()}},
\code{\link[=rw_range]{rw_range()}},
\code{\link[=skewness_vec]{skewness_vec()}}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Vector Function}
