% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_attributes}
\alias{get_attributes}
\title{Get Attributes}
\usage{
get_attributes(.data)
}
\arguments{
\item{.data}{An R object from which attributes are to be extracted.}
}
\value{
A list of attributes of the input R object, excluding row.names.
}
\description{
The \code{get_attributes} function takes an R object as input and
returns its attributes, omitting the row.names attribute.
}
\details{
This function retrieves the attributes of a given R object,
excluding the row.names attribute.
}
\examples{
get_attributes(rw30())
get_attributes(iris)
get_attributes(mtcars)

}
\seealso{
Other Utility Functions: 
\code{\link{confidence_interval}()},
\code{\link{convert_snake_to_title_case}()},
\code{\link{generate_caption}()},
\code{\link{rand_walk_column_names}()},
\code{\link{rand_walk_helper}()},
\code{\link{running_quantile}()},
\code{\link{std_cum_max_augment}()},
\code{\link{std_cum_mean_augment}()},
\code{\link{std_cum_min_augment}()},
\code{\link{std_cum_prod_augment}()},
\code{\link{std_cum_sum_augment}()},
\code{\link{subset_walks}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Utility Functions}
