% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_cummax_init}
\alias{rcpp_cummax_init}
\title{Fast Cumulative Maximum with Initial Value}
\usage{
rcpp_cummax_init(x, initial_value = 0)
}
\arguments{
\item{x}{A numeric vector}

\item{initial_value}{A numeric value to add to all cumulative maximums}
}
\value{
A numeric vector of cumulative maximums
}
\description{
This function computes the cumulative maximum of a numeric vector with an initial value offset.
Implemented in C++ for improved performance.
}
