% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_cummean_init}
\alias{rcpp_cummean_init}
\title{Fast Cumulative Mean with Initial Value}
\usage{
rcpp_cummean_init(x, initial_value = 0)
}
\arguments{
\item{x}{A numeric vector}

\item{initial_value}{A numeric value to add to all cumulative means}
}
\value{
A numeric vector of cumulative means
}
\description{
This function computes the cumulative mean of a numeric vector with an initial value offset.
Implemented in C++ for improved performance.
}
