% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_cummin_init}
\alias{rcpp_cummin_init}
\title{Fast Cumulative Minimum with Initial Value}
\usage{
rcpp_cummin_init(x, initial_value = 0)
}
\arguments{
\item{x}{A numeric vector}

\item{initial_value}{A numeric value to add to all cumulative minimums}
}
\value{
A numeric vector of cumulative minimums
}
\description{
This function computes the cumulative minimum of a numeric vector with an initial value offset.
Implemented in C++ for improved performance.
}
