% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_cumprod_init}
\alias{rcpp_cumprod_init}
\title{Fast Cumulative Product with Initial Value}
\usage{
rcpp_cumprod_init(x, initial_value = 1)
}
\arguments{
\item{x}{A numeric vector}

\item{initial_value}{A numeric value to multiply with cumulative products}
}
\value{
A numeric vector of cumulative products
}
\description{
This function computes the cumulative product of (1 + x) multiplied by initial value.
Implemented in C++ for improved performance.
}
