% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_cumsum_init}
\alias{rcpp_cumsum_init}
\title{Fast Cumulative Sum with Initial Value}
\usage{
rcpp_cumsum_init(x, initial_value = 0)
}
\arguments{
\item{x}{A numeric vector}

\item{initial_value}{A numeric value to add to all cumulative sums}
}
\value{
A numeric vector of cumulative sums
}
\description{
This function computes the cumulative sum of a numeric vector with an initial value offset.
Implemented in C++ for improved performance.
}
