% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-cumulative-functions.R
\name{skewness_vec}
\alias{skewness_vec}
\title{Compute Skewness of a Vector}
\usage{
skewness_vec(.x)
}
\arguments{
\item{.x}{A numeric vector of length four or more.}
}
\value{
The skewness of a vector
}
\description{
This function takes in a vector as it's input and will return the skewness
of that vector. The length of this vector must be at least four numbers. The
skewness explains the 'tailedness' of the distribution of data.

\deqn{\frac{\frac{1}{n} \sum (x - \mu)^3}{\left(\frac{1}{n} \sum (x - \mu)^2\right)^{3/2}}}
}
\details{
A function to return the skewness of a vector.
}
\examples{
set.seed(123)
skewness_vec(rnorm(100, 3, 2))

}
\seealso{
\url{https://en.wikipedia.org/wiki/Skewness}

Other Vector Function:
\code{\link[=cgmean]{cgmean()}},
\code{\link[=chmean]{chmean()}},
\code{\link[=ckurtosis]{ckurtosis()}},
\code{\link[=cmean]{cmean()}},
\code{\link[=cmedian]{cmedian()}},
\code{\link[=crange]{crange()}},
\code{\link[=csd]{csd()}},
\code{\link[=cskewness]{cskewness()}},
\code{\link[=cvar]{cvar()}},
\code{\link[=euclidean_distance]{euclidean_distance()}},
\code{\link[=kurtosis_vec]{kurtosis_vec()}},
\code{\link[=rw_range]{rw_range()}}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Vector Function}
