% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-functions.R
\name{cgmean}
\alias{cgmean}
\title{Cumulative Geometric Mean}
\usage{
cgmean(.x)
}
\arguments{
\item{.x}{A numeric vector}
}
\value{
A numeric vector
}
\description{
A function to return the cumulative geometric mean of a vector.
}
\details{
A function to return the cumulative geometric mean of a vector.
\code{exp(cummean(log(.x)))}
}
\examples{
x <- mtcars$mpg

cgmean(x)

}
\seealso{
Other Vector Function: 
\code{\link{chmean}()},
\code{\link{ckurtosis}()},
\code{\link{cmean}()},
\code{\link{cmedian}()},
\code{\link{csd}()},
\code{\link{cskewness}()},
\code{\link{cvar}()},
\code{\link{kurtosis_vec}()},
\code{\link{name}},
\code{\link{rw_range}()},
\code{\link{skewness_vec}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Vector Function}
