% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-functions.R
\name{cmedian}
\alias{cmedian}
\title{Cumulative Median}
\usage{
cmedian(.x)
}
\arguments{
\item{.x}{A numeric vector}
}
\value{
A numeric vector
}
\description{
A function to return the cumulative median of a vector.
}
\details{
A function to return the cumulative median of a vector.
}
\examples{
x <- mtcars$mpg

cmedian(x)

}
\seealso{
Other Vector Function: 
\code{\link{cgmean}()},
\code{\link{chmean}()},
\code{\link{ckurtosis}()},
\code{\link{cmean}()},
\code{\link{csd}()},
\code{\link{cskewness}()},
\code{\link{cvar}()},
\code{\link{kurtosis_vec}()},
\code{\link{name}},
\code{\link{rw_range}()},
\code{\link{skewness_vec}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Vector Function}
