% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-functions.R
\name{crange}
\alias{crange}
\title{Cumulative Range}
\usage{
crange(.x)
}
\arguments{
\item{.x}{A numeric vector}
}
\value{
A numeric vector
}
\description{
A function to return the cumulative range of a vector.
}
\details{
A function to return the cumulative range of a vector. It uses \code{max(.x[1:k]) - min(.x[1:k])} as
the basis of the function.
}
\examples{

x <- mtcars$mpg

crange(x)

}
\seealso{
Other Vector Function: 
\code{\link{cgmean}()},
\code{\link{chmean}()},
\code{\link{ckurtosis}()},
\code{\link{cmean}()},
\code{\link{cmedian}()},
\code{\link{csd}()},
\code{\link{cskewness}()},
\code{\link{cvar}()},
\code{\link{kurtosis_vec}()},
\code{\link{rw_range}()},
\code{\link{skewness_vec}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Vector Function}
