% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-geom-brown-motion.R
\name{geometric_brownian_motion}
\alias{geometric_brownian_motion}
\title{Geometric Brownian Motion}
\usage{
geometric_brownian_motion(
  .num_walks = 25,
  .n = 100,
  .mu = 0,
  .sigma = 0.1,
  .initial_value = 100,
  .delta_time = 0.003,
  .return_tibble = TRUE
)
}
\arguments{
\item{.num_walks}{Total number of simulations.}

\item{.n}{Total time of the simulation, how many \code{n} points in time.}

\item{.mu}{Expected return}

\item{.sigma}{Volatility}

\item{.initial_value}{Integer representing the initial value.}

\item{.delta_time}{Time step size.}

\item{.return_tibble}{The default is TRUE. If set to FALSE then an object
of class matrix will be returned.}
}
\value{
A tibble/matrix
}
\description{
Create a Geometric Brownian Motion.
}
\details{
Geometric Brownian Motion (GBM) is a statistical method for modeling
the evolution of a given financial asset over time. It is a type of stochastic
process, which means that it is a system that undergoes random changes over
time.

GBM is widely used in the field of finance to model the behavior of stock
prices, foreign exchange rates, and other financial assets. It is based on
the assumption that the asset's price follows a random walk, meaning that it
is influenced by a number of unpredictable factors such as market trends,
news events, and investor sentiment.

The equation for GBM is:

\if{html}{\out{<div class="sourceCode">}}\preformatted{ dS/S = mdt + sdW
}\if{html}{\out{</div>}}

where S is the price of the asset, t is time, m is the expected return on the
asset, s is the volatility of the asset, and dW is a small random change in
the asset's price.

GBM can be used to estimate the likelihood of different outcomes for a given
asset, and it is often used in conjunction with other statistical methods to
make more accurate predictions about the future performance of an asset.

This function provides the ability of simulating and estimating the parameters
of a GBM process. It can be used to analyze the behavior of financial
assets and to make informed investment decisions.
}
\examples{
library(ggplot2)

geometric_brownian_motion()

geometric_brownian_motion() |>
  ggplot(aes(x = x, y = y, group = walk_number, color = walk_number)) +
  geom_line() +
  labs(title = "Geometric Brownian Motion", x = "Time", y = "Value") +
  theme_minimal() +
  theme(legend.position = "none")

}
\seealso{
Other Generator Functions: 
\code{\link{brownian_motion}()},
\code{\link{random_normal_drift_walk}()},
\code{\link{random_normal_walk}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Generator Functions}
