% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-uniform-walk.R
\name{random_uniform_walk}
\alias{random_uniform_walk}
\title{Generate Multiple Random Uniform Walks in Multiple Dimensions}
\usage{
random_uniform_walk(
  .num_walks = 25,
  .n = 100,
  .min = 0,
  .max = 1,
  .initial_value = 0,
  .samp = TRUE,
  .replace = TRUE,
  .sample_size = 0.8,
  .dimensions = 1
)
}
\arguments{
\item{.num_walks}{An integer specifying the number of random walks to generate. Default is 25.}

\item{.n}{An integer specifying the number of steps in each walk. Default is 100.}

\item{.min}{A numeric value indicating the minimum of the uniform distribution. Default is 0.}

\item{.max}{A numeric value indicating the maximum of the uniform distribution. Default is 1.}

\item{.initial_value}{A numeric value indicating the initial value of the walks. Default is 0.}

\item{.samp}{A logical value indicating whether to sample the uniform distribution values. Default is TRUE.}

\item{.replace}{A logical value indicating whether sampling is with replacement. Default is TRUE.}

\item{.sample_size}{A numeric value between 0 and 1 specifying the proportion of \code{.n} to sample. Default is 0.8.}

\item{.dimensions}{An integer specifying the number of dimensions (1, 2, or 3). Default is 1.}
}
\value{
A tibble containing the generated random walks with columns depending on the number of dimensions:
\itemize{
\item \code{walk_number}: Factor representing the walk number.
\item \code{step_number}: Step index.
\item \code{y}: If \code{.dimensions = 1}, the value of the walk at each step.
\item \code{x}, \code{y}: If \code{.dimensions = 2}, the values of the walk in two dimensions.
\item \code{x}, \code{y}, \code{z}: If \code{.dimensions = 3}, the values of the walk in three dimensions.
}

The following are also returned based upon how many dimensions there are and could be any of x, y and or z:
\itemize{
\item \code{walk_number}: Factor representing the walk number.
\item \code{x}: Step index.
\item \code{y}: Uniform distribution values.
\item \code{cum_sum}: Cumulative sum of \code{y}.
\item \code{cum_prod}: Cumulative product of \code{y}.
\item \code{cum_min}: Cumulative minimum of \code{y}.
\item \code{cum_max}: Cumulative maximum of \code{y}.
}

The tibble includes attributes for the function parameters.
}
\description{
The \code{random_uniform_walk} function generates multiple random walks in 1, 2, or 3 dimensions.
Each walk is a sequence of steps where each step is a random draw from a uniform distribution.
The user can specify the number of walks, the number of steps in each walk, and the
parameters of the uniform distribution (min and max). The function
also allows for sampling a proportion of the steps and optionally sampling with replacement.
}
\examples{
set.seed(123)
random_uniform_walk()

set.seed(123)
random_uniform_walk(.dimensions = 3) |>
  head() |>
  t()

}
\seealso{
Other Generator Functions: 
\code{\link{brownian_motion}()},
\code{\link{discrete_walk}()},
\code{\link{geometric_brownian_motion}()},
\code{\link{random_normal_drift_walk}()},
\code{\link{random_normal_walk}()},
\code{\link{random_weibull_walk}()},
\code{\link{random_wilcox_walk}()}

Other Continuous Distribution: 
\code{\link{brownian_motion}()},
\code{\link{geometric_brownian_motion}()},
\code{\link{random_normal_drift_walk}()},
\code{\link{random_normal_walk}()},
\code{\link{random_weibull_walk}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Continuous Distribution}
\concept{Generator Functions}
