% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-random-weibull-walk.R
\name{random_weibull_walk}
\alias{random_weibull_walk}
\title{Generate Multiple Random Weibull Walks in Multiple Dimensions}
\usage{
random_weibull_walk(
  .num_walks = 25,
  .n = 100,
  .shape = 1,
  .scale = 1,
  .initial_value = 0,
  .samp = TRUE,
  .replace = TRUE,
  .sample_size = 0.8,
  .dimensions = 1
)
}
\arguments{
\item{.num_walks}{Integer. Number of walks to generate. Default is 25.}

\item{.n}{Integer. Number of steps in each walk. Default is 100.}

\item{.shape}{Numeric. Shape parameter of the Weibull distribution. Default is 1.}

\item{.scale}{Numeric. Scale parameter of the Weibull distribution. Default is 1.}

\item{.initial_value}{Numeric. Starting value of the walk. Default is 0.}

\item{.samp}{Logical. Whether to sample the steps. Default is TRUE.}

\item{.replace}{Logical. Whether sampling is with replacement. Default is TRUE.}

\item{.sample_size}{Numeric. Proportion of steps to sample (0-1). Default is 0.8.}

\item{.dimensions}{Integer. Number of dimensions (1, 2, or 3). Default is 1.}
}
\value{
A data frame with the random walks and cumulative statistics as columns.

A tibble containing the generated random walks with columns depending
on the number of dimensions:
\itemize{
\item \code{walk_number}: Factor representing the walk number.
\item \code{step_number}: Step index.
\item \code{y}: If \code{.dimensions = 1}, the value of the walk at each step.
\item \code{x}, \code{y}: If \code{.dimensions = 2}, the values of the walk in two dimensions.
\item \code{x}, \code{y}, \code{z}: If \code{.dimensions = 3}, the values of the walk in three dimensions.
}

The following are also returned based upon how many dimensions there are and
could be any of x, y and or z:
\itemize{
\item \code{cum_sum}: Cumulative sum of \code{dplyr::all_of(.dimensions)}.
\item \code{cum_prod}: Cumulative product of \code{dplyr::all_of(.dimensions)}.
\item \code{cum_min}: Cumulative minimum of \code{dplyr::all_of(.dimensions)}.
\item \code{cum_max}: Cumulative maximum of \code{dplyr::all_of(.dimensions)}.
\item \code{cum_mean}: Cumulative mean of \code{dplyr::all_of(.dimensions)}.
}
}
\description{
A Weibull random walk is a stochastic process in which each step is drawn
from the Weibull distribution, a flexible distribution commonly used to model
lifetimes, reliability, and extreme values. This function allows for the
simulation of multiple independent random walks in one, two, or three
dimensions, with user control over the number of walks, steps, and the shape
and scale parameters of the Weibull distribution. Sampling options allow for
further customization, including the ability to sample a proportion of steps
and to sample with or without replacement. The resulting data frame includes
cumulative statistics for each walk, making it suitable for simulation
studies and visualization.
}
\details{
The \code{random_weibull_walk} function generates multiple random walks in 1, 2, or 3 dimensions.
Each walk is a sequence of steps where each step is a random draw from the
Weibull distribution using \code{stats::rweibull()}. The user can specify the
number of walks, the number of steps in each walk, and the parameters
\code{.shape} and \code{.scale} for the Weibull distribution. The function also allows
for sampling a proportion of the steps and optionally sampling with replacement.
}
\examples{
set.seed(123)
random_weibull_walk()

set.seed(123)
random_weibull_walk(.dimensions = 3) |>
   head() |>
   t()

}
\seealso{
Other Generator Functions: 
\code{\link{brownian_motion}()},
\code{\link{discrete_walk}()},
\code{\link{geometric_brownian_motion}()},
\code{\link{random_normal_drift_walk}()},
\code{\link{random_normal_walk}()},
\code{\link{random_smirnov_walk}()},
\code{\link{random_t_walk}()},
\code{\link{random_uniform_walk}()},
\code{\link{random_wilcox_walk}()}

Other Continuous Distribution: 
\code{\link{brownian_motion}()},
\code{\link{geometric_brownian_motion}()},
\code{\link{random_normal_drift_walk}()},
\code{\link{random_normal_walk}()},
\code{\link{random_t_walk}()},
\code{\link{random_uniform_walk}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Continuous Distribution}
\concept{Generator Functions}
